requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
scikit-learn>=1.3
xgboost>=2.0
streamlit>=1.28
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9
rapidfuzz>=3.5
pytest>=7.4
python-dotenv>=1.0
